requirements.txt
flask
requests
websockets
python-dotenv
pyaudio
onnxruntime
Unidecode
flask-cors